@@ -16,10 +16,10 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 0.3.14+11
+version: 0.3.15+14
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.7.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,6 +30,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations:
+    sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -37,8 +39,7 @@ dependencies:
   audioplayers: ^6.0.0
   gap: ^3.0.1
   dart_random_choice: ^0.2.0
-  dart_date: ^1.1.1
-  permission_handler: ^11.0.0
+  # permission_handler: ^11.0.0
   flex_color_scheme: ^8.1.0
   intl: ^0.19.0
   auto_size_text: ^3.0.0
@@ -53,14 +54,9 @@ dependencies:
   google_fonts: ^6.1.0
   flutter_launcher_icons: ^0.14.1
   syncfusion_flutter_sliders: ^28.1.41
-  # cloud_text_to_speech: ^1.1.3
   cloud_text_to_speech:
     git:
       url: https://github.com/morbo12/cloud_text_to_speech_flutter
-    # path: ../cloud_text_to_speech
-  # horizontal_center_date_picker:
-    # git:
-      # url: https://lars0280@dev.azure.com/lars0280/Soundboard/_git/flutter_horizontal_center_date_picker
   archive: ^4.0.2
   flutter_volume_controller: ^1.3.1
   flutter_toastr: ^1.0.3
@@ -68,28 +64,17 @@ dependencies:
   syncfusion_flutter_core: ^28.1.41
   riverpod_annotation: ^2.3.3
   flutter_riverpod: ^2.4.9
-  easy_localization: ^3.0.4
   file_picker: ^9.0.2
-  # camera_windows: ^0.2.6+1
-  # camera: ^0.11.0+2
   animated_introduction: ^1.0.4
   shared_preferences: ^2.5.2
   
-  # _fe_analyzer_shared: ^74.0.0
-  # analyzer: ^6.9.0
-
-# dev_dependencies:
-#   msix: ^3.16.7
-
-#   flutter_test:
-#     # sdk: flutter
-#   flutter_lints: ^4.0.0
-#   riverpod_generator: ^2.3.9
-#   build_runner: ^2.4.8
-#   custom_lint: ^0.6.2
-#   riverpod_lint: ^2.3.7
-#   analyzer: ^6.5.2
-#   meta: '>1.5.0'
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  integration_test:
+    sdk: flutter
+  mockito: ^5.4.5
+  build_runner: ^2.4.15
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
