@@ -1,4 +1,5 @@
 import 'package:flex_color_scheme/flex_color_scheme.dart';
+import 'package:soundboard/utils/logger.dart';
 
 class AppThemes {
   // Keep the original constants for reference
@@ -24,16 +25,19 @@ class AppThemes {
     purpleBrown,
 
     // Then add all remaining schemes
-    ...FlexScheme.values.where((scheme) => ![
-          greyLaw,
-          aquaBlue,
-          ebonyClay,
-          outerSpace,
-          blueWhale,
-          sanJuanBlue,
-          blueM3,
-          purpleBrown,
-        ].contains(scheme)),
+    ...FlexScheme.values.where(
+      (scheme) =>
+          ![
+            greyLaw,
+            aquaBlue,
+            ebonyClay,
+            outerSpace,
+            blueWhale,
+            sanJuanBlue,
+            blueM3,
+            purpleBrown,
+          ].contains(scheme),
+    ),
   ];
 
   static String getName(FlexScheme scheme) {
@@ -102,12 +106,13 @@ class AppThemes {
     }
 
     // Convert camelCase to words with spaces
-    final result = name
-        .replaceAllMapped(
-          RegExp(r'([A-Z])'),
-          (match) => ' ${match.group(0)}',
-        )
-        .trim();
+    final result =
+        name
+            .replaceAllMapped(
+              RegExp(r'([A-Z])'),
+              (match) => ' ${match.group(0)}',
+            )
+            .trim();
 
     // Capitalize first letter
     if (result.isNotEmpty) {
@@ -173,8 +178,9 @@ class AppThemes {
         return scheme;
       }
     }
+    const Logger logger = Logger('AppThemes');
+    logger.e("ERROR ${name.toLowerCase()}");
 
-    print("ERROR ${name.toLowerCase()}");
     throw ArgumentError('Unknown theme name: $name');
   }
 
