@@ -1,107 +1,158 @@
-import 'package:flutter/foundation.dart';
+// ssml_penalty_event.dart
+import 'dart:math';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_toastr/flutter_toastr.dart';
-import 'package:intl/intl.dart';
-import 'package:soundboard/constants/globals.dart';
-import 'package:soundboard/constants/providers.dart';
-import 'package:soundboard/features/cloud_text_to_speech/providers.dart';
-import 'package:soundboard/features/screen_home/presentation/live/data/class_penalty_type.dart';
 import 'package:soundboard/features/innebandy_api/data/class_match_event.dart';
-import 'package:soundboard/features/innebandy_api/data/class_match.dart';
+import 'package:soundboard/features/screen_home/presentation/live/data/class_penalty_type.dart';
 import 'package:soundboard/properties.dart';
+import 'class_ssml_base.dart';
 
-class SsmlPenaltyEvent {
-  final IbyMatchEvent matchEvent;
-  final NumberFormat formatter = NumberFormat("0");
-  final WidgetRef ref;
+class PenaltyTemplates {
+  static final List<
+    String Function(
+      String time,
+      String period,
+      String number,
+      String name,
+      String team,
+      String penalty,
+    )
+  >
+  templates = [
+    // Standard format
+    (time, period, number, name, team, penalty) => '''
+      Nummer $number, $name 
+      i $team utvisas $penalty. 
+      Tid: <say-as interpret-as='duration' format='ms'>$time</say-as>
+    ''',
 
-  SsmlPenaltyEvent({
-    required this.ref,
-    required this.matchEvent,
-  });
-  // data.matchTeamName == selectedMatch.awayTeam
-  String whosEvent() {
-    final selectedMatch = ref.read(selectedMatchProvider);
-    return (matchEvent.matchTeamName == selectedMatch.homeTeam)
-        ? "hemmalaget"
-        : "bortalaget";
-  }
+    // Time-first format
+    (time, period, number, name, team, penalty) => '''
+      Vid <say-as interpret-as='duration' format='ms'>$time</say-as> i $period
+      utvisas nummer $number, $name 
+      i $team för $penalty
+    ''',
 
-  String homeOrAwayScore() {
-    String score = "";
-    if (whosEvent() == "bortalaget") {
-      score = "${matchEvent.goalsAwayTeam} - ${matchEvent.goalsHomeTeam}";
-    } else {
-      score = "${matchEvent.goalsHomeTeam} - ${matchEvent.goalsAwayTeam}";
-    }
-
-    // if (matchEvent.goalsAwayTeam)
-    return score;
-  }
-
-  String penaltyName() {
-    Map<String, String> penaltyInfo =
-        PenaltyTypes.getPenaltyInfo(matchEvent.penaltyCode);
-    if (kDebugMode) {
-      print("penaltyName: ${penaltyInfo['time']} för ${penaltyInfo['name']}");
-    }
-    String penaltyString = "";
-    if (penaltyInfo['time'] != "Unknown") {
-      penaltyString =
-          "${penaltyInfo['time']} för <break time='200ms' /> ${penaltyInfo['name']}";
-    }
-    return penaltyString;
-  }
-
-  // Nummer z i <laget> utvisas x minuter för <penaltyname>
-  // String who = whoScored(matchTeamName: matchEvent.matchTeamName);
-  // Kvitterar
-  // 1-0 till hemmalaget – målskytt nr 12 Pelle Karlsson – assist nr 11 Kent Persson och 9 Sven Hansson – Tid 11:04.
-  // <lag> <händelse> med <1-0,0-1>
-  String whatWasTheTime() {
-    return matchEvent.minute != 0
-        ? "${formatter.format(matchEvent.minute)}:${formatter.format(matchEvent.second)}"
-        : "${formatter.format(matchEvent.second)} sekunder";
-  }
-
-  String stripTeamSuffix(String teamName) {
-    return teamName.replaceAll(RegExp(r' \([A-Z]\)'), '');
-  }
-
-  Future<bool> getSay(BuildContext context) async {
-    // String say = matchEvent.matchTeamName;
-
-    String say =
-        "Nummer ${matchEvent.playerShirtNo}, ${matchEvent.playerName} i ${stripTeamSuffix(matchEvent.matchTeamName)} utvisas ${penaltyName()}. Tid: <say-as interpret-as='duration' format='ms'>${whatWasTheTime()}</say-as> ";
-
-    if (kDebugMode) {
-      print("SAY: $say");
-    }
-    FlutterToastr.show(say, context,
-        duration: FlutterToastr.lengthLong,
-        position: FlutterToastr.bottom,
-        backgroundColor: Colors.black,
-        textStyle: const TextStyle(color: Colors.white));
-    final textToSpeechService = ref.read(textToSpeechServiceProvider);
-    final ssml = await textToSpeechService.getTtsNoFile(text: say);
-    ref.read(azCharCountProvider.notifier).state += say.length;
-    SettingsBox().azCharCount +=
-        say.length; // TODO: Should check if getTts was successful
-
-    // await eventAudioPlayer.setVolume(1.0);
-    await jingleManager.audioManager
-        .playBytes(audio: ssml.audio.buffer.asUint8List(), ref: ref);
-    // eventAudioPlayer.stop();
-    // eventAudioPlayer.release();
-
-    return true;
-  }
-
-// Nybro IF tar ledningen med 1-0. Målskytt utan assistans nummer 24 Peter Eriksson. Tid 12.14
-
-  List<String> goalSays = [
-    "<hemmalag> utökar ledningen till x-y, mål av <person>",
-    "Mål av <person>, <lag> leder med x-y",
+    // Team-focused format
+    (time, period, number, name, team, penalty) => '''
+      $team får en utvisning. 
+      $penalty på nummer $number, $name.
+      Tid: <say-as interpret-as='duration' format='ms'>$time</say-as>
+    ''',
   ];
+}
+
+class SsmlPenaltyEvent extends BaseSsmlEvent {
+  final IbyMatchEvent matchEvent;
+  final Random _random = Random();
+
+  SsmlPenaltyEvent({required super.ref, required this.matchEvent})
+    : super(loggerName: 'SsmlPenaltyEvent');
+
+  @override
+  String formatAnnouncement() {
+    _validateEventData();
+    return _formatPenaltyAnnouncement();
+  }
+
+  void _validateEventData() {
+    if (matchEvent.playerShirtNo == null || matchEvent.playerShirtNo! <= 0) {
+      throw ValidationException('Invalid player number');
+    }
+    if (matchEvent.playerName.trim().isEmpty) {
+      throw ValidationException('Player name is required');
+    }
+    if (matchEvent.matchTeamName.trim().isEmpty) {
+      throw ValidationException('Team name is required');
+    }
+  }
+
+  String _formatPenaltyAnnouncement() {
+    final time = _formatTime();
+    final teamName = stripTeamSuffix(matchEvent.matchTeamName);
+    final penaltyInfo = _getPenaltyInfo();
+
+    final templateIndex = _random.nextInt(PenaltyTemplates.templates.length);
+    final template = PenaltyTemplates.templates[templateIndex];
+
+    final announcement = template(
+      time,
+      matchEvent.periodName,
+      matchEvent.playerShirtNo.toString(),
+      matchEvent.playerName,
+      teamName,
+      penaltyInfo,
+    );
+
+    return _addProsodyVariation(
+      _addRandomPauses(announcement.trim().replaceAll(RegExp(r'\s+'), ' ')),
+    );
+  }
+
+  String _formatTime() {
+    if (matchEvent.minute == 0) {
+      return formatter.format(matchEvent.second);
+    }
+    return '${formatter.format(matchEvent.minute)}:${formatter.format(matchEvent.second)}';
+  }
+
+  String _getPenaltyInfo() {
+    final penaltyInfo = PenaltyTypes.getPenaltyInfo(matchEvent.penaltyCode);
+    if (penaltyInfo['time'] == "Unknown") {
+      return penaltyInfo['name'] ?? 'okänd utvisning';
+    }
+    return "${penaltyInfo['time']} för ${penaltyInfo['name']}";
+  }
+
+  String _addProsodyVariation(String text) {
+    final rates = ['slow', 'medium', 'fast'];
+    final pitches = ['low', 'medium', 'high'];
+
+    return wrapWithProsody(
+      text,
+      rate: rates[_random.nextInt(rates.length)],
+      pitch: pitches[_random.nextInt(pitches.length)],
+    );
+  }
+
+  String _addRandomPauses(String text) {
+    if (!text.contains('.')) return text;
+
+    return text
+        .split('.')
+        .where((s) => s.trim().isNotEmpty)
+        .map((sentence) {
+          final pauseLength = _random.nextInt(300) + 200;
+          return '${sentence.trim()}. <break time="${pauseLength}ms"/>';
+        })
+        .join(' ');
+  }
+
+  @override
+  Future<bool> getSay(BuildContext context) async {
+    try {
+      final announcement = formatAnnouncement();
+      logger.d("Announcement: $announcement");
+
+      await showToast(context, announcement);
+      await playAnnouncement(announcement);
+
+      return true;
+    } catch (e, stackTrace) {
+      logger.e('Failed to process penalty announcement', e, stackTrace);
+      await showToast(
+        context,
+        "Ett fel uppstod vid utvisningsannonsering",
+        isError: true,
+      );
+      return false;
+    }
+  }
+}
+
+class ValidationException implements Exception {
+  final String message;
+  ValidationException(this.message);
+
+  @override
+  String toString() => 'ValidationException: $message';
 }
