@@ -1,13 +1,15 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 import 'package:soundboard/common_widgets/button.dart';
 import 'package:soundboard/constants/globals.dart';
 import 'package:soundboard/features/screen_home/application/audioplayer/audioplayer_func.dart';
+import 'package:soundboard/utils/logger.dart';
 
 class StopGoalRow extends ConsumerWidget {
   const StopGoalRow({super.key});
+
+  final Logger logger = const Logger('StopGoalRow');
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -17,9 +19,7 @@ class StopGoalRow extends ConsumerWidget {
         Button(
           noLines: 1,
           onTap: () {
-            if (kDebugMode) {
-              print("[StopGoalRow] STOP");
-            } // playerStop();
+            logger.d("STOP was pressed");
             jingleManager.audioManager.stopAll(ref);
           },
           primaryText: 'STOP',
@@ -30,6 +30,7 @@ class StopGoalRow extends ConsumerWidget {
         Button(
           noLines: 1,
           onTap: () {
+            logger.d("GOAL was pressed");
             playGoal2(ref);
           },
           primaryText: 'MÅL',
