@@ -1,44 +1,17 @@
-import 'dart:io';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:soundboard/common_widgets/button.dart';
+import 'package:soundboard/features/screen_settings/presentation/widgets/class_cache_service.dart';
 
 class CleanCacheButton extends StatefulWidget {
   const CleanCacheButton({super.key}); // Updated constructor
 
   @override
-  CleanCacheButtonToDirState createState() => CleanCacheButtonToDirState();
+  CleanCacheButtonState createState() => CleanCacheButtonState();
 }
 
-class CleanCacheButtonToDirState extends State<CleanCacheButton> {
-  // File? file;
-  // final ValueNotifier<String?> selectedPath = ValueNotifier(null);
-
-  Future<void> _deleteCacheDirectory() async {
-    try {
-      // Get the cache directory
-      final Directory cacheDir = await getApplicationCacheDirectory();
-
-      // Check if the directory exists
-      if (await cacheDir.exists()) {
-        // Delete the directory and its contents
-        await cacheDir.delete(recursive: true);
-
-        if (kDebugMode) {
-          print("Cache directory deleted: ${cacheDir.path}");
-        }
-      } else {
-        if (kDebugMode) {
-          print("Cache directory does not exist: ${cacheDir.path}");
-        }
-      }
-    } catch (e) {
-      if (kDebugMode) {
-        print("Error deleting cache directory: $e");
-      }
-    }
-  }
+class CleanCacheButtonState extends State<CleanCacheButton> {
+  final CacheService _cacheService = CacheService();
+  bool _isLoading = false;
 
   @override
   Widget build(BuildContext context) {
@@ -48,50 +21,148 @@ class CleanCacheButtonToDirState extends State<CleanCacheButton> {
           style: ElevatedButton.styleFrom(
             foregroundColor: Theme.of(context).colorScheme.onErrorContainer,
             backgroundColor: Theme.of(context).colorScheme.errorContainer,
-            fixedSize: Size.fromHeight(100),
-            // side: BorderSide(
-            // width: 1, color: Theme.of(context).colorScheme.primaryContainer),
+            fixedSize: const Size.fromHeight(100),
           ),
           noLines: 1,
           isSelected: true,
-          onTap: () async {
-            final Directory _cacheDir = await getApplicationCacheDirectory();
-
-            // Invoke the file picker UI function
-            bool? confirm = await showDialog<bool>(
-              context: context,
-              builder: (BuildContext context) {
-                return AlertDialog(
-                  title: Text('Confirm Deletion'),
-                  content: Text(
-                      'Are you sure you want to delete the cache? \n${_cacheDir.path}'),
-                  actions: <Widget>[
-                    TextButton(
-                      child: Text('Cancel'),
-                      onPressed: () {
-                        Navigator.of(context).pop(false); // Return false
-                      },
-                    ),
-                    TextButton(
-                      child: Text('Delete'),
-                      onPressed: () {
-                        Navigator.of(context).pop(true); // Return true
-                      },
-                    ),
-                  ],
-                );
-              },
-            );
-
-            if (confirm == true) {
-              await _deleteCacheDirectory();
-            }
-          },
+          onTap: _isLoading ? null : () => _handleCacheDeletion(),
           secondaryText: 'N/A',
-          primaryText:
-              "!!! DANGER - Delete jingle cache - DANGER !!!", // Use directory name for the button text
+          primaryText: "!!! DANGER - Delete jingle cache - DANGER !!!",
         ),
       ],
+    );
+  }
+
+  // Note: No BuildContext parameter here
+  Future<void> _handleCacheDeletion() async {
+    setState(() {
+      _isLoading = true;
+    });
+
+    try {
+      // Get cache directory and size
+      final cacheDir = await _cacheService.getCacheDirectory();
+      final cacheSize = await _cacheService.calculateCacheSize(cacheDir);
+      final formattedSize = _cacheService.formatBytes(cacheSize);
+
+      if (!mounted) return;
+
+      setState(() {
+        _isLoading = false;
+      });
+
+      // Now we can safely use context since we're back in sync code
+      final shouldDelete =
+          await _showConfirmationDialog(cacheDir.path, formattedSize);
+
+      if (shouldDelete == true) {
+        _showLoadingIndicator();
+
+        final success = await _cacheService.clearCache(cacheDir);
+
+        if (!mounted) return;
+
+        // Close loading dialog
+        Navigator.of(context).pop();
+
+        // Show feedback
+        _showResultFeedback(success);
+      }
+    } catch (e) {
+      if (!mounted) return;
+
+      setState(() {
+        _isLoading = false;
+      });
+
+      _showErrorFeedback(e.toString());
+    }
+  }
+
+  // Helper methods that use the current context
+  Future<bool?> _showConfirmationDialog(String cachePath, String cacheSize) {
+    return showDialog<bool>(
+      context: context, // Using the current context is safe here
+      builder: (BuildContext dialogContext) {
+        return AlertDialog(
+          title: const Text('Confirm Cache Deletion'),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              const Text('Are you sure you want to delete the cache?'),
+              const SizedBox(height: 16),
+              Text('Location: $cachePath'),
+              const SizedBox(height: 8),
+              Text('Size: $cacheSize'),
+              const SizedBox(height: 16),
+              const Text(
+                'Warning: This will remove all cached jingles.',
+                style: TextStyle(color: Colors.orange),
+              ),
+            ],
+          ),
+          actions: <Widget>[
+            TextButton(
+              child: const Text('Cancel'),
+              onPressed: () {
+                Navigator.of(dialogContext).pop(false);
+              },
+            ),
+            TextButton(
+              style: TextButton.styleFrom(
+                foregroundColor: Colors.red,
+              ),
+              child: const Text('Delete'),
+              onPressed: () {
+                Navigator.of(dialogContext).pop(true);
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  void _showLoadingIndicator() {
+    showDialog(
+      context: context, // Using the current context is safe here
+      barrierDismissible: false,
+      builder: (BuildContext dialogContext) {
+        return AlertDialog(
+          content: Row(
+            children: [
+              CircularProgressIndicator(
+                color: Theme.of(context).colorScheme.primary,
+              ),
+              const SizedBox(width: 20),
+              const Text("Clearing cache...")
+            ],
+          ),
+        );
+      },
+    );
+  }
+
+  void _showResultFeedback(bool success) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(success
+            ? 'Cache successfully cleared'
+            : 'Cache directory not found or already empty'),
+        backgroundColor: success ? Colors.green : Colors.orange,
+        duration: const Duration(seconds: 3),
+      ),
+    );
+  }
+
+  void _showErrorFeedback(String errorMessage) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text('Failed to clear cache: $errorMessage'),
+        backgroundColor: Colors.red,
+        duration: const Duration(seconds: 5),
+      ),
     );
   }
 }
