@@ -1,12 +1,11 @@
 import 'dart:io';
-
 import 'package:archive/archive_io.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:soundboard/common_widgets/button.dart';
 import 'package:soundboard/constants/globals.dart';
 import 'package:soundboard/features/screen_settings/presentation/widgets/file_picker_util.dart';
+import 'package:soundboard/utils/logger.dart';
 
 class UploadButtonAll extends StatefulWidget {
   const UploadButtonAll({super.key}); // Updated constructor
@@ -18,7 +17,7 @@ class UploadButtonAll extends StatefulWidget {
 class UploadButtonToDirState extends State<UploadButtonAll> {
   File? file;
   final ValueNotifier<String?> selectedPath = ValueNotifier(null);
-
+  final Logger logger = const Logger('UploadButtonAll');
   Future<void> _unzipFile({required String? file}) async {
     final Directory appSupportDir = await getApplicationCacheDirectory();
     final Directory targetDir =
@@ -29,17 +28,13 @@ class UploadButtonToDirState extends State<UploadButtonAll> {
           recursive: true); // Ensure the target directory exists
     }
 
-    if (kDebugMode) {
-      print("Extracting files to $targetDir");
-    }
+    logger.d("Extracting files to $targetDir");
 
     try {
       await extractFileToDisk(
           file!, targetDir.path); // Updated to use target directory
     } catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
+      logger.d(e.toString());
     }
   }
 
