@@ -1,17 +1,21 @@
 import 'dart:io';
-import 'package:flutter/foundation.dart';
+import 'dart:math';
 import 'package:flutter/material.dart';
-import 'package:file_picker/file_picker.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:soundboard/common_widgets/button.dart';
 import 'package:soundboard/constants/globals.dart';
+import 'package:soundboard/features/jingle_manager/application/class_audiocategory.dart';
+import 'package:soundboard/features/jingle_manager/application/class_jingle_manager.dart';
 import 'package:soundboard/features/screen_settings/presentation/widgets/file_picker_util.dart';
+import 'package:soundboard/utils/logger.dart';
 
 class UploadButtonToDir extends StatefulWidget {
   final String directoryName; // Updated to be more descriptive
 
-  const UploadButtonToDir(
-      {super.key, required this.directoryName}); // Updated constructor
+  const UploadButtonToDir({
+    super.key,
+    required this.directoryName,
+  }); // Updated constructor
 
   @override
   UploadButtonToDirState createState() => UploadButtonToDirState();
@@ -20,62 +24,46 @@ class UploadButtonToDir extends StatefulWidget {
 class UploadButtonToDirState extends State<UploadButtonToDir> {
   File? file;
   final ValueNotifier<String?> selectedPath = ValueNotifier(null);
-
-  // Future<void> _unzipFile({required String? file}) async {
-  //   final Directory appSupportDir = await getApplicationCacheDirectory();
-  //   final Directory targetDir = Directory(
-  //       '${appSupportDir.path}/${widget.directoryName}'); // Create target directory
-
-  //   if (!await targetDir.exists()) {
-  //     await targetDir.create(
-  //         recursive: true); // Ensure the target directory exists
-  //   }
-
-  //   if (kDebugMode) {
-  //     print("Extracting files to $targetDir");
-  //   }
-
-  //   try {
-  //     await extractFileToDisk(
-  //         file!, targetDir.path); // Updated to use target directory
-  //   } catch (e) {
-  //     if (kDebugMode) {
-  //       print(e);
-  //     }
-  //   }
-  // }
+  final Logger logger = const Logger('UploadButtonToDir');
 
   Future<void> _copyFileToDestination(List<File>? files) async {
     if (files == null) return;
 
     final Directory appSupportDir = await getApplicationCacheDirectory();
-    final Directory targetDir =
-        Directory('${appSupportDir.path}/${widget.directoryName}');
+    final Directory targetDir = Directory(
+      '${appSupportDir.path}/${widget.directoryName}',
+    );
 
     if (!await targetDir.exists()) {
       await targetDir.create(recursive: true);
     }
     for (var sourceFile in files) {
       // final File sourceFile = File(file);
-      final String targetPath = Platform.isWindows
-          ? '${targetDir.path}/${sourceFile.path.split('\\').last}'
-          : '${targetDir.path}/${sourceFile.path.split('/').last}';
+      final String targetPath =
+          Platform.isWindows
+              ? '${targetDir.path}/${sourceFile.path.split('\\').last}'
+              : '${targetDir.path}/${sourceFile.path.split('/').last}';
 
       try {
         await sourceFile.copy(targetPath);
-        if (kDebugMode) {
-          print("File copied to $targetPath");
-        }
+        logger.d("File copied to $targetPath");
       } catch (e) {
-        if (kDebugMode) {
-          print("Failed to copy file: $e");
-        }
+        logger.d("Failed to copy file: $e");
       }
     }
   }
 
   @override
   Widget build(BuildContext context) {
+    final tmp_category =
+        widget.directoryName == "GenericJingles"
+            ? AudioCategory.genericJingle
+            : widget.directoryName == "GoalJingles"
+            ? AudioCategory.goalJingle
+            : widget.directoryName == "ClapJingles"
+            ? AudioCategory.clapJingle
+            : AudioCategory.awayTeamJingle;
+
     return Button(
       style: ElevatedButton.styleFrom(
         foregroundColor: Theme.of(context).colorScheme.onTertiaryContainer,
@@ -94,9 +82,7 @@ class UploadButtonToDirState extends State<UploadButtonToDir> {
           onMultipleFilesSelected: (files) async {
             if (!mounted) return;
 
-            if (kDebugMode) {
-              print("VALUE: $files");
-            }
+            logger.d("VALUE: $files");
 
             // Copy the files to destination
             await _copyFileToDestination(files);
@@ -111,7 +97,9 @@ class UploadButtonToDirState extends State<UploadButtonToDir> {
           },
         );
       },
-      secondaryText: 'N/A',
+      secondaryText:
+          "(${jingleManager.audioManager.audioInstances.where((element) => element.audioCategory == tmp_category).length.toString()})",
+
       primaryText:
           widget.directoryName, // Use directory name for the button text
     );
